address comments + rename file
